feat: add custom hook, move handlers from app to useHandlers
